Changed hebi_allocid typedef to an integer.
This fixes CERT/MISRA issues with casting pointers to/from integers.
Unfortunately, we no longer have type-checking on the opaque handle
type, but this brings us into CERT/MISRA compliance.

+++ src/z/zdestroyn.c
@@ -20,6 +20,6 @@ hebi_zdestroyn(size_t count, struct hebi_integer r[count])
 		r[i].hz_resv = 0;
 		r[i].hz_used = 0;
 		r[i].hz_sign = 0;
-		r[i].hz_allocid = (int)(intptr_t)HEBI_ALLOC_INVALID;
+		r[i].hz_allocid = HEBI_ALLOC_INVALID;
 	}
 }
