+++ src/z/zdestroyv.c
@@ -22,6 +22,6 @@ hebi_zdestroyv(size_t count, const hebi_zptr r[count])
 		r[i]->hz_resv = 0;
 		r[i]->hz_used = 0;
 		r[i]->hz_sign = 0;
-		r[i]->hz_allocid = (int)(intptr_t)HEBI_ALLOC_INVALID;
+		r[i]->hz_allocid = HEBI_ALLOC_INVALID;
 	}
 }
